mlflow

+++ src/mlops_train.py
@@ -80,7 +80,6 @@ def train_and_evaluate(config_path):
         #         "r2": r2
         #     }
         #     json.dump(scores, f)
-        
         # with open(params_file, "w") as f:
         #     params = {
         #         "alpha": alpha,
@@ -96,9 +95,13 @@ def train_and_evaluate(config_path):
 
         # model_path = config["model_path"]
         # joblib.dump(lr, model_path)
-        os.makedirs(model_dir, exist_ok=True)
-        model_path = os.path.join(model_dir, "model.joblib")
-        joblib.dump(lr, model_path)
+        if os.path.isfile(model_dir):  # Check if a file with the same name exists
+            os.remove(model_dir)  # Delete the file
+
+        if not os.path.exists(model_dir):  # Now safely create the directory
+            os.makedirs(model_dir, exist_ok=True)
+            model_path = os.path.join(model_dir, "model.joblib")
+            joblib.dump(lr, model_path)
 
 
 if __name__=="__main__":
